streamlit/requirements.txt
streamlit>=1.23.0
numpy>=1.20.0
langchain>=0.0.350
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
langchain-community>=0.0.10